Relocates launchpadSaml to endpoint directory.
This makes more sense as it contains, you know, endpoints for saml
authorization.
