+++ packages/api/tests/endpoints/fixtures/_certificateFixture.js
