+++ packages/api/tests/endpoints/fixtures/bad-metadata.xml
