+++ packages/api/tests/endpoints/fixtures/launchpad-sbx-metadata.xml
